refactor: 라이프사이클 조회 서비스 분리
