+++ src/main/java/umc/cockple/demo/domain/exercise/controller/ExerciseLifecycleController.java
@@ -10,7 +10,7 @@ import umc.cockple.demo.domain.exercise.dto.ExerciseDeleteDTO;
 import umc.cockple.demo.domain.exercise.dto.ExerciseDetailDTO;
 import umc.cockple.demo.domain.exercise.dto.ExerciseEditDetailDTO;
 import umc.cockple.demo.domain.exercise.dto.ExerciseUpdateDTO;
-import umc.cockple.demo.domain.exercise.service.ExerciseQueryService;
+import umc.cockple.demo.domain.exercise.service.query.ExerciseLifecycleQueryService;
 import umc.cockple.demo.domain.exercise.service.command.ExerciseCommandService;
 import umc.cockple.demo.global.response.BaseResponse;
 import umc.cockple.demo.global.response.code.status.CommonSuccessCode;
@@ -22,7 +22,7 @@ import umc.cockple.demo.global.security.utils.SecurityUtil;
 public class ExerciseLifecycleController implements ExerciseLifecycleApi {
 
     private final ExerciseCommandService exerciseCommandService;
-    private final ExerciseQueryService exerciseQueryService;
+    private final ExerciseLifecycleQueryService exerciseLifecycleQueryService;
 
     @Override
     public ResponseEntity<BaseResponse<ExerciseCreateDTO.Response>> createExercise(
@@ -60,7 +60,7 @@ public class ExerciseLifecycleController implements ExerciseLifecycleApi {
     public ResponseEntity<BaseResponse<ExerciseDetailDTO.Response>> getExerciseDetail(Long exerciseId) {
         Long memberId = SecurityUtil.getCurrentMemberId();
 
-        ExerciseDetailDTO.Response response = exerciseQueryService.getExerciseDetail(
+        ExerciseDetailDTO.Response response = exerciseLifecycleQueryService.getExerciseDetail(
                 exerciseId, memberId);
 
         return BaseResponse.of(CommonSuccessCode.OK, response);
@@ -69,7 +69,7 @@ public class ExerciseLifecycleController implements ExerciseLifecycleApi {
     @Override
     public ResponseEntity<BaseResponse<ExerciseEditDetailDTO.Response>> getExerciseForEdit(Long exerciseId) {
         Long memberId = SecurityUtil.getCurrentMemberId();
-        ExerciseEditDetailDTO.Response response = exerciseQueryService.getExerciseForEdit(exerciseId, memberId);
+        ExerciseEditDetailDTO.Response response = exerciseLifecycleQueryService.getExerciseForEdit(exerciseId, memberId);
         return BaseResponse.of(CommonSuccessCode.OK, response);
     }
 }
